Workaround: Don't highlight claim IDs
The initial hash in claim IDs in the input were being highlighted as the
non-overlapping claim.

The fix would be to defined a different named syntax for each window.

The workaround is to note that all claims are at least 2 inches wide, so
the non-overlapping claim must have at least two hashes in a row, which
the IDs obviously don't.