Refactoring: Avoid @state synthetic variable
It wasn't a good name for a variable anyway. Now that partition_by is
dividing up a list, just pass the list straight to it.

Unfortunately this makes $modulus an even worse name for a variable, since
it's now not even the modulus any more: there's no longer a spot between
the map and the partitioning to tweak it, so the tweaking happens in the
partitioning.