Refactoring: Calculate $offset separately
Rename $modulus to $offset, since it holds the amount by which frequencies
are offset between cycles.

Calculate it first simply, using sum. It's effectively another loop, but
it's so much clearer this way, rather than trying to bodge it into the map
that calculates all the running frequencies.

It also means the final change, not needed in the list of running totals,
can simply be pop-ed off, which is more readable than the hash slice.

In the map, simply use $freq, a new state variable, to remember the
current frequency between iterations.

We now have two variables which are never referred to by name: their only
use is the place they are declared — but I think Perl insists on their
having names anyway.