Refactoring: Co-ordinates as hashes
Have each co-ordinate be represented by x and y fields, rather than
elements 0 and 1 of an array.

Numbers are better when iterating over each dimension in turn, but here
they are nested, and it's clearer to refer to them by name.

The + 0 is irksome, though; without it there are warnings and the wrong
answer, as the tieing in %+ somehow goes awry. Is minmax using regexp?