Bug Fix: Ignore infinite areas
This was happening to get the right answer with both the demo and real
input. But theoretically the largest area inside the min–max bounds could
extend the bounds, and therefore be technically infinite — but with its
area inside the bounds be larger than any finite areas.

Copy from the other attempt at Part 1 having %area_size be a hash that
gets removed from as areas become ineligible.

Because the boundary check is now on the boundary (rather than past it),
the check is equality for both min and max — which means that we don't
actually need to know which is which, and any can be used. (The other any
to the other program, because these are both bounds of the same
dimension.)

The // is a bit unfortunate, but alternatives involve widening the scope
of $area_id or not being able to use each @coord (or having to fiddle with
$area_id to make it match the first loop's).

Or, actually ...