Refactoring: Avoid state ++ // 0 awkwardness
Switch to reading the input with while <> and push, rather than map <>.

In this algorithm, co-ordinates don't need to know their own IDs, so the
push can be done first, then the highest index from the array used. That's
by definition the same as will be returned by the each on that array
later, so is much less icky all round, even allowing for while push hiding
the data flow.