Enhancement: Stack goes down, not up
Add to the stack of numbers of metadata items at the bottom of the buffer,
so it goes down the page, rather than after the input line — which was
adding to the top of the list, pushing each of the existing numbers down
the page.

This has the advantage of reducing flicker and being more obvious what's
going on: once added to the stack, a number stays at that position on the
screen until it is reached.

Also, avoids the need for the extra blank line at the bottom to cope with
the difference in dd's behaviour in terms of where it leaves the cursor
after deleting the final line of the buffer and any other line: we're now
always deleting the final line, so it's consistent without any extra
effort.