Bug Fix: Only count s/^/oo/ once
When adding extra "o"s, the /g was supposed to make the pattern match up
to once at each of the beginning and the end.

But it was matching twice at the beginning on a string that didn't
previously have any "o"s in it. I'm not sure why, but splitting it into
two separate substitutions means the /g isn't needed, which avoids the
problem.

It also means that (?{...}) isn't needed.