Bug Fix: Let patterns match in all positions
While \K provides an equivalent to (?<=...) from the point of view of an
individual match, it isn't equivalent enough in this case. Specifically,
it was gobbling up the two characters to the left, advancing the match
position, and preventing those characters from ever matching in the middle
of a rule.

Now all rules match exactly one character, so the match can proceed over
each character in turn.