Solved: Part  1
Change the approach: rather than storing the direction symbolically,
convert each to an axis and a delta along that axis.

Then reaching corners and junctions involves modifying one or both of
those.

Position is now {pos}[0, 1] rather then {x},{y}, so that the axis can
easily be flipped arithmetically.

Handily, / always involve flipping the sign on the delta, and \ never does
(which I'm taking as a sign this a good design).

The junction turning behaviours are encoded 0 .. 2, such that 1 is left, 2
is straight on, and 0 is right. That gives the property that the axis
needs switching if left than 2, and the delta sign needs flipping if the
axis just switched to has the same number as the behaviour.

A bit obscure, but the behaviours need encoding somehow.