Bug Fix: Process carts in correct order
The *_by functions don't work as I'd hoped, being able to return a list of
fields to try in turn (as sort does in Perl 6), so to get a numeric sort I
need to combine them into a single number.

Multiplying the y co-ordinate by the list works, but is awkward, so
instead switch to Sort::Key::Multi, which does what I want here.

Unfortunately it requires telling how many fields there are, so has the
slightly awkward name ii_keysort, but it does the right thing.

This was happening to get the right answer for my input; presumably sort
is stable, and in the crashing tick carts were being processed by x
co-ordinate — the same — and then in the order found in the file — which
happened to be right. (On non-crashing ticks, it doesn't matter.)

Found by a forthcoming change which switches carts to being a hash rather
than an array. That's in random order, and for my input gave the wrong
answer about half the time.