Solved: Part 2
A cart's ID also needs to be used in its data, so the current cart knows
its own ID. (Else iterate over keys rather than values, and add another
level of nesting on the look-ups._

But this does mean that %cart_pos can be populated at the same time.

Now have this holding IDs. Also, delete rather than just blanking out (I
think I could've done this before).

A cart may crash into one which hasn't moved yet in this tick, so will be
deleted from %cart before it has been iterated to, so have the check at
the start of the loop that it still exists.

I would've expected that ii_keysort would take a copy of all the values,
so $_ would, in its turn, still end up set to the hash-ref. But actually
Perl complains “Use of freed value in iteration at”. So use @sorted to
ensure they are definitely copied.