Enhancement: Leave carts on the track diagram
Don't bother updating the track diagram with straight pieces underneath
the initial positions of the carts.

The only things that are checked for are curves and junctions; carts
always continue in their current direction, presuming track to be there
for them — so the initial carts don't get in the way of anything, and
there's no need to remove them.