Abandoned: Recursion is the wrong approach
This approach should get there in the end, but it's doing a depth-first
search, which is silly for finding the shortest path.

Abandon this and work out how to do what I presume is a breadth-first
search ... but commit what I've done so far, in case it does ever turn out
to be useful.