Refactoring: 1D map
Everything's in reading order. Vertical movements can be made by
adding/subtracting the map width. A commenter on Reddit used a 1D grid in
a previous day's puzzle this year.

Clearly it's much less hassle simply to do this in 1D.

Pretty much every line has changed in this diff, but they are
line-for-line replacements, and the braces help to match up the old and
new.