Refactoring: Don't hard-code the letters
Nothing else now depends on the letters ‘E’ and ‘G’, so remove them from
the parsing, to make this entirely generic (as well as a couple of
characters shorter).

Mainly because it's slightly amusing to have an answer that doesn't
feature these tokens at all. It also means it would work for more than 2
unit types, so long as all units were fighting all other types.