Workaround: Quell DDP
DDP was emitting undef warnings when $\ was set to undef (maybe it sets $/
to it or something?). Shut it up by local-izing the change to just the
operation that requires it.

(This also saves a line, though I think it adds to the ‘line noise’, and
is arguably more awkward to read, other than as an atomic Perl idiom.)