Solved: Part 1
I actually solved it yesterday, but only ran it with Exoji2e's input
rather than my own, so got the wrong answer!

They key was this comment:

  # XXX consider looping over units, not positions

I meant that as an elegance thing, but actually it was needed to get the
right answer. Just tracking positions could be problematic if a unit moved
into a later reading-order position from which a unit had been eliminated.

For instance with this state:

  .G.
  GEE

Suppose the G at the top eliminates the E directly below it, later in the
reading order. Then the G on the left can move into the square where the E
used to be, and attack the E at the right.

Then the iteration over positions reaches the now-eliminated E. It checks
whether there is actually still a unit there — and there is: the G that
it's just moved there in the previous iteration! So that G gets an
incorrect second go at attacking the remaining E during this round.

To actually track when units have been eliminated, we need to iterate over
units, rather than positions. Store a unit's position in the unit (as well
as its key), so that we have handy access to it for checking moves from.

It's a little unfortunate to store this twice, while and each doesn't work
well with sorting, so this seems the cleanest way.

Remove no-longer-needed debugging output. Though maybe it will be needed
if I want to track down the differences in positions between mine and
Gusto Bravo's for some rounds.