Source: Better function name: next_pos()
nearest_target() was finding the nearest target, but it was actually
returning the next position (may be the same as the current one), so have
its name reflect that.

Now it's no longer recursive, and no longer returns early from the middle
of some nested loops, it probably no longer needs to be a function anyway;
it may simplify things to inline it.