Refactoring: Inline next_pos()
next_pos() was only called once, and passing in a big list of parameters
made that quite a hassle. Simply inline what it was doing, with the
variables we have anyway.

This also makes it clearer to see which variables are hashes, because they
aren't referred to via scalar references.

And it fixes a bug: some of the function was referring to the main
program's variable %unit, not the parameter $unit. Because \%unit is
passed to $unit, they are actually the same thing, but it still
shouldn't've been doing that.