Solved: Part 1
Could use a couple of conditional lasts to give up iterating through the
ops when either we've already found 3 matches or there aren't enough ops
left to find 3, but this is so quick anyway, it doesn't seem worth it. And
not having those checks makes the code cleaner and simpler.