Enhancement: Don't slurp the program
There's a program at the end, which we need.

Handily, it's separated from the samples by 3 empty lines. So if we
initially split on specifically 1 empty line (rather than Perl's default
‘paragraph mode’ of any number of empty lines) then the first record not
matching a sample will be an empty line.

At which point, the program can then be read in, a line at a time. So
switch back to normal $\ behaviour for that, by localizing the other
behaviour in a block.

View with -w.