Progress: Work out ops for each number
Don't actually need the op code for each number; simply going straight
from the number to the op sub itself will do. We end up with a hash where
the keys are mixture of op codes and numbers, but they don't overlap in
form, so that doesn't matter.

Repeatedly looking for a code with only one matching number is sufficient
for my input; there's no need to also check for numbers that only match
one code.