Throwaway: Trying slice_grep
The previous throwaway was using grep with foreach instead of a while with
a next unless in it. It saved a line, but overall seemed more complicated.

I also looked for a Cpan module for hash partition, but couldn't find one.

slice_grep allows for grep-ing a hash and returning both parts, which is
slightly simpler than just using grep on the keys and having to get the
value afterwards.

Except that it still involves looking the value up inside the grep block —
the hard-to-read $_{$_} (which itself involves the entire hash being
copied), and then using each on the result requires referencing and
immediately de-referencing the list that's returned into a hash. And we
still have the awkwardness of turning a value of a reference to a
single-element array into the that element.

So overall, it seems more complicated.

Also, I didn't actually get it working. But now the complexity is such, it
doesn't seem worth pursuing. Saving here as an idea, before reverting.