Refactoring: Use pairfirst
_This_ is what I was looking for when trying out grep (in 39eb9b25) and
slice_grep (in e04148ee): pairfirst.

The trick is not to try to grab _all_ the matching opcodes; simply getting
any one (so, the first) will do; the outer while loop will take care of
reaching the rest in good time.

Inspired by Drbagy on Reddit, who uses grep, but only assigning to one
element (so is effectively using first):
https://old.reddit.com/r/adventofcode/comments/a6mf8a/2018_day_16_solutions/#ebwcy13