Refactoring: Explicit eof check
While rare in Perl, an explicit eof check here seems to be the
least-clunky way of avoiding the warning when the final undef from <>
being matched against the pattern.

De-activating all uninitialized warnings wasn't clear where, and why, they
were de-activated, and potentially also hides a useful warning elsewhere.

A defined check involves also an explicit assignment to $_ and lots of
parens, which obscures what's going on.

And <> // '' works but is non-obvious and a lot of punctuation (admittedly
in nice neat pairs) to inflict on readers.